Convert Code to use `<cstdint>`

+++ SVM/functions/math.h
@@ -7,49 +7,49 @@
 
 namespace SVM::math
 {
-    inline unsigned long long next_number()
+    inline uint64_t next_number()
     {
-        const auto num = SVM::Globals::program_stack.top<unsigned long long>();
+        const auto num = SVM::Globals::program_stack.top<uint64_t>();
         SVM::Globals::program_stack.pop();
         return num;
     }
 
 	inline void ADD()
     {
-        SVM::Globals::program_stack.push<unsigned long long>(next_number() + next_number(), 8);
+        SVM::Globals::program_stack.push<uint64_t>(next_number() + next_number(), 8);
     }
 
 	inline void SUB()
     {
-        const unsigned long long num1 = next_number();
-        const unsigned long long num2 = next_number();
-        SVM::Globals::program_stack.push<unsigned long long>(num1 - num2, 8);
+        const uint64_t num1 = next_number();
+        const uint64_t num2 = next_number();
+        SVM::Globals::program_stack.push<uint64_t>(num1 - num2, 8);
     }
 
 	inline void MUL()
     {
-        SVM::Globals::program_stack.push<unsigned long long>(next_number() * next_number(), 8);
+        SVM::Globals::program_stack.push<uint64_t>(next_number() * next_number(), 8);
     }
 
 	inline void DIV()
     {
-        const unsigned long long num1 = next_number();
-        const unsigned long long num2 = next_number();
-        SVM::Globals::program_stack.push<unsigned long long>(num1 / num2, 8);
+        const uint64_t num1 = next_number();
+        const uint64_t num2 = next_number();
+        SVM::Globals::program_stack.push<uint64_t>(num1 / num2, 8);
     }
 
 	inline void MOD()
     {
-        const unsigned long long num1 = next_number();
-        const unsigned long long num2 = next_number();
-        SVM::Globals::program_stack.push<unsigned long long>(num1 % num2, 8);
+        const uint64_t num1 = next_number();
+        const uint64_t num2 = next_number();
+        SVM::Globals::program_stack.push<uint64_t>(num1 % num2, 8);
     }
 
 	inline void POW()
     {
-        const unsigned long long number = next_number();
-        const unsigned long long power = next_number();
-        SVM::Globals::program_stack.push<unsigned long long>(static_cast<unsigned long long>(pow(number, power)), 8);
+        const uint64_t number = next_number();
+        const uint64_t power = next_number();
+        SVM::Globals::program_stack.push<uint64_t>(static_cast<uint64_t>(pow(number, power)), 8);
     }
 }
